+++ DocuVault/LoginPage.xaml.cs
@@ -11,7 +11,7 @@ namespace DocuVault
 {
     public partial class LoginPage : Page
     {
-        private readonly AccessDB _accessDB;
+        private readonly UserService _usersService;
         private bool isPasswordVisible = false;
         private string _password; // Store the password here
         private bool _isSyncingPassword = false; // Flag to prevent recursive updates
@@ -19,7 +19,8 @@ namespace DocuVault
         public LoginPage()
         {
             InitializeComponent();
-            _accessDB = new AccessDB(); // Initialize AccessDB
+            var accessDB = new AccessDB(); // Initialize AccessDB
+            _usersService = new UserService(accessDB); // Initialize UsersService
         }
 
         private void Btn_SignIn_Click(object sender, RoutedEventArgs e)
@@ -28,63 +29,15 @@ namespace DocuVault
             string password = _password;
 
             // Authenticate the user
-            if (IsAuthenticated(email, password))
+            if (_usersService.IsAuthenticated(email, password))
             {
-                UserService user = GetUserFromDatabase(email);
+                UserService user = _usersService.GetUserFromDatabase(email);
                 this.NavigationService.Navigate(new DashboardPage(user));
             }
             else
             {
                 MessageBox.Show("Invalid login credentials");
             }
-        }
-
-        private bool IsAuthenticated(string email, string password)
-        {
-            return _accessDB.Execute(connection =>
-            {
-                string hashedPassword = HashPassword(password);
-
-                string query = "SELECT COUNT(*) FROM Users WHERE Email = @Email AND Password = @Password";
-                using (OleDbCommand command = new OleDbCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@Email", email);
-                    command.Parameters.AddWithValue("@Password", hashedPassword);
-
-                    return (int)command.ExecuteScalar() > 0;
-                }
-            });
-        }
-
-        private string HashPassword(string password)
-        {
-            using (SHA256 sha256 = SHA256.Create())
-            {
-                byte[] hashBytes = sha256.ComputeHash(Encoding.UTF8.GetBytes(password));
-                return Convert.ToBase64String(hashBytes);
-            }
-        }
-
-        private UserService GetUserFromDatabase(string email)
-        {
-            return _accessDB.Execute(connection =>
-            {
-                string query = "SELECT Email, IsAdmin FROM Users WHERE Email = @Email";
-                using (OleDbCommand command = new OleDbCommand(query, connection))
-                {
-                    command.Parameters.AddWithValue("@Email", email);
-
-                    using (OleDbDataReader reader = command.ExecuteReader())
-                    {
-                        if (reader.Read())
-                        {
-                            bool isAdmin = reader.GetBoolean(reader.GetOrdinal("IsAdmin"));
-                            return new UserService(email, isAdmin);
-                        }
-                        return null;
-                    }
-                }
-            });
         }
 
         private void TextBox_Email_TextChanged(object sender, TextChangedEventArgs e)
